Skip if inactive.

+++ app/workers/entry_deleter.rb
@@ -3,9 +3,13 @@ class EntryDeleter
   sidekiq_options queue: :worker_slow
 
   def perform(feed_id)
-    feed = Feed.find(feed_id)
-    unless feed.protected
-      delete_entries(feed_id)
+    if Subscription.where(feed_id: feed_id, active: true).exists?
+      feed = Feed.find(feed_id)
+      if !feed.protected && feed.subscriptions_count > 0
+        delete_entries(feed_id)
+      end
+    else
+      Librato.increment('entry.destroy_skip')
     end
   end
 
@@ -16,7 +20,6 @@ class EntryDeleter
       entries_to_keep = Entry.where(feed_id: feed_id).order('published DESC').limit(entry_limit).pluck('entries.id')
       entries_to_delete = Entry.select(:id, :public_id).where(feed_id: feed_id, starred_entries_count: 0).where.not(id: entries_to_keep)
       entries_to_delete_ids = entries_to_delete.map {|entry| entry.id }
-      entries_to_delete_public_ids = entries_to_delete.map {|entry| entry.public_id }
 
       # Delete records
       UnreadEntry.where(entry_id: entries_to_delete_ids).delete_all
@@ -24,9 +27,9 @@ class EntryDeleter
       RecentlyReadEntry.where(entry_id: entries_to_delete_ids).delete_all
       entries_to_delete.delete_all
 
-      key_created_at = FeedbinUtils.redis_feed_entries_created_at_key(feed_id)
-      key_published = FeedbinUtils.redis_feed_entries_published_key(feed_id)
       if entries_to_delete_ids.present?
+        key_created_at = FeedbinUtils.redis_feed_entries_created_at_key(feed_id)
+        key_published = FeedbinUtils.redis_feed_entries_published_key(feed_id)
         SearchIndexRemove.perform_async(entries_to_delete_ids)
         $redis[:sorted_entries].zrem(key_created_at, entries_to_delete_ids)
         $redis[:sorted_entries].zrem(key_published, entries_to_delete_ids)
